SVEA-40: Moved configuration and controller to window._svea
Before the configuration was set on window directly by ssn.phtml.
Unfortunately the controller was initialized before the script block in
ssn/phtml was executed so all values that _SveaController.initialize
tried to read was undefined. This passed because of javascript but it
also meant that the controller initialization was broken.